+++ .github/workflows/release.yaml
@@ -27,6 +27,8 @@ jobs:
       run: |
         sudo apt-get update
         sudo apt-get install -y libgtk-3-dev webkit2gtk-4.0 libappindicator3-dev librsvg2-dev patchelf
+    - name: ui tests
+      run: yarn test
     - name: install app dependencies and build it
       run: yarn && yarn tauri build
     - uses: tauri-apps/tauri-action@v0
